+++ system/platform-core/src/main/java/org/platformlambda/core/serializers/SimpleMapper.java
@@ -34,8 +34,6 @@ public class SimpleMapper {
     private static final Logger log = LoggerFactory.getLogger(SimpleMapper.class);
 
     private static final String SNAKE_CASE_SERIALIZATION = "snake.case.serialization";
-    private static final Set<String> safeModels = new HashSet<>();
-    private static final String[] SAFE_GROUPS = {"java.util.", "java.lang."};
     private final SimpleObjectMapper mapper;
     private final SimpleObjectMapper snakeMapper;
     private final SimpleObjectMapper camelMapper;
@@ -51,18 +49,6 @@ public class SimpleMapper {
         this.mapper = new SimpleObjectMapper(getJson(snake, true), getJson(snake, false));
         this.snakeMapper = new SimpleObjectMapper(getJson(true, true), getJson(true, false));
         this.camelMapper = new SimpleObjectMapper(getJson(false, true), getJson(false, false));
-        /*
-         * Optionally, load white list for authorized PoJo
-         */
-        AppConfigReader reader = AppConfigReader.getInstance();
-        String models = reader.getProperty("safe.data.models");
-        if (models != null) {
-            List<String> list = Utility.getInstance().split(models, ", ");
-            for (String m: list) {
-                safeModels.add(m.endsWith(".") ? m : m +".");
-            }
-            log.info("Safe data models {}", safeModels);
-        }
     }
 
     public Gson getJson(boolean snake, boolean isMap) {
@@ -136,43 +122,6 @@ public class SimpleMapper {
      */
     public SimpleObjectMapper getCamelCaseMapper() {
         return camelMapper;
-    }
-
-    public SimpleObjectMapper getSafeMapper(Class<?> cls) {
-        return getSafeMapper(cls.getTypeName());
-    }
-
-    public SimpleObjectMapper getSafeMapper(String clsName) {
-        if (permittedDataModel(clsName)) {
-            return mapper;
-        } else {
-            throw new IllegalArgumentException("Class "+clsName+" not in safe.data.models");
-        }
-    }
-
-    private boolean permittedDataModel(String clsName) {
-        // accept all types if safe.data.models feature is not enabled
-        if (safeModels.isEmpty()) {
-            // feature not enabled
-            return true;
-        }
-        // always allow primitive types including byte[]
-        if (!clsName.contains(".")) {
-            return true;
-        }
-        // accept safe java.util and java.lang classes
-        for (String m: SAFE_GROUPS) {
-            if (clsName.startsWith(m)) {
-                return true;
-            }
-        }
-        // validate with white list
-        for (String m: safeModels) {
-            if (clsName.startsWith(m)) {
-                return true;
-            }
-        }
-        return false;
     }
 
     /// Custom serializers ///